Move minutely probe to their own file and module
Move the minutely probes to their own files and module so it's not
defined in the separate hook file and those files are can grow very long
now.

Move the tests to its own file so I can keep an overview of what is what
easier there.

Move it to its own module so that we can move other probes there too and
we know where we can find and define them in the future.

We can decide to create aliases for the previous constant naming of
these probes, but I assume people aren't interfacing with them directly,
as they are private APIs inside the AppSignal gem.
